+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat(logging): cloudwatch logs account setting for subscription filter
 - feat(backoff): added escape hatch to configure number of retries and starting backoff
 - feat(kinesis): allow customization of retention, streaming mode and shard count
+- feat(logging): allow users to customize lambda processor for firehose
 
 ### Fixed